+++ SDK V3.X/Bot/BotLibrary/Model/Credentials.cs
@@ -2,38 +2,38 @@
 
 namespace BotLibrary.Model
 {
-    public class BotCredentials
+    public class Credentials
     {
         public ServerConfig ServerConfig { get; set; }
 
         public UserConfig UserConfig { get; set; }
 
-        public BotCredentials()
+        public Credentials()
         {
             ServerConfig = new();
             UserConfig = new();
         }
 
-        public static Boolean FromJsonNode(JSONNode jsonNode, out BotCredentials botCredentials)
+        public static Boolean FromJsonNode(JSONNode jsonNode, out Credentials credentials)
         {
             if (jsonNode is not null)
             {
-                botCredentials = new ();
+                credentials = new ();
 
                 if (UserConfig.FromJsonNode(jsonNode["userConfig"], out UserConfig userconfig))
-                    botCredentials.UserConfig = userconfig;
+                    credentials.UserConfig = userconfig;
                 else
                     return false;
 
                 if (ServerConfig.FromJsonNode(jsonNode["serverConfig"], out ServerConfig serverConfig))
-                    botCredentials.ServerConfig = serverConfig;
+                    credentials.ServerConfig = serverConfig;
                 else
                     return false;
 
                 return true;
             }
             else
-                botCredentials = new();
+                credentials = new();
 
             return false;
         }